Added check for NPE on proxy key.

+++ branches/changemessaging/flexLibrary/flex-src/net/digitalprimates/persistence/hibernate/HibernateBean.as
@@ -35,7 +35,10 @@ package net.digitalprimates.persistence.hibernate
 		}
 		
 		public function set proxyKey( value:Object ):void {
-			StateRepository.removeFromNewEntityList( __proxyKey.toString() );
+			if (__proxyKey)
+			{
+				StateRepository.removeFromNewEntityList( __proxyKey.toString() );
+			}
 			__proxyKey = value;
 		}
